requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
confluent-kafka>=2.3.0
psycopg2-binary>=2.9.9
tabulate>=0.9.0